dev-requirements.txt
--editable .
responses>=0.5.0
djangorestframework-simplejwt>=5.3.1
django-allauth>=65.5.0
django-user-agents>=0.4.0
djangorestframework>=3.15.2
flake8
requests
requests-oauthlib
sphinx
furo
django
